Remove generate-site script, prefer HTTP server configuration
Instead of providing the “generate-site” script, which used symlinks to
trick the HTTP server into serving the /releases path from a different
location, add a note in the README file advising to configure an alias
instead. Having symlinks in the generated _site/ directory is confusing
for Jekyll, and letting the HTTP server handle the /releases prefix is
more robust because it doesn't depend on stray files under _site/ being
left untouched by Jekyll.