Add microdata markup for posts
This is a first pass at adding schema.org Microdata markup to pages
which use the "post" template. This adds inline metadata to the HTML
markup which helps e.g. crawlers have a better view of content.

An introduction to Microdata can be found here:

  https://developer.mozilla.org/en-US/docs/Web/HTML/Microdata

The reference documentation for the metadata schemas can be found at:

  https://schema.org