Add some improvements to the hardware compatibility tables (#78)
* hw: i.MX53 and i.MX51 use the same GPU, list the latter as supported

Note that these two entries still have empty field that need filling.
Probably the driver is Mesa's freedreno because the Imageon Z460 is
the original branding for Adreno 200, but needs checking.

* hw: List i.MX6 as supported with the RDK backend

This combination is mentioned as supported in
https://github.com/Igalia/meta-webkit/issues/217

* hw: Remove USE_BACKEND_ prefix from RDK backend entries

The prefix is always the same, so it adds little information. The
important part of the option is the rest of the name, so leave it
at that.