Use a local copy of the Source Sans 3 variable font
The CSS specifies local() a few times with names the font may already
be installed as, including using Source Sans Pro as fallback, to avoid
loading the fonts over the wire where possible. Note that Source Sans 3
is a slightly more refined version of Source Sans Pro, so it seems like
a reasonable option to use it rather than the original. Plus, it comes
as a variable font, which allows downloading only two font files for all
possible weight combinations.

Closes #97