Editorial modifications and extra explanations to make it more readable
Some parts were needlessly complex or hard to read without adding much
value; shortening them with equivalent phrases probably helps.

The subsection entitled "Browsers, engines and projects" is likely
"...and ports", because half of the section is spent discussing WebKit
ports and "projects" is used as a roughly equivalent of "engines".

Finally added some text like a paragraph to try to address directly
the questions of what an embedded browser is, and not only explain
that it's built along an embedded system; and add examples of other
WebKit ports to contrast them with WPE.