Remove section about Raspbian/RPiOS images
In recent meetings (e.g. WebKit Codecamp Oct 2022) it was decided to
drop them, as they are unfavourable compared to the Yocto images,
specially in terms of showing performance.

Besides, some parts of the text are clearly out of date, like
mentioning $45 as the price of a RPi 3, which is currently not even
possible to obtain due to supply shortages.