11ty: Upgrade to version 2.0 🎉
Version 2 has been released a while ago, comes with speed improvements,
and is lighter on dependencies. What's not to like? From a review of the
release notes [1] it looks like there is no need to take care for any of
the breaking changes, and the upgrade helper plug-in [2] passed with
flying colors.

[1] https://www.11ty.dev/blog/eleventy-v2/
[2] https://github.com/11ty/eleventy-upgrade-help