11ty: Replace usage of "url" filter with HTMLBase plugin
Use the new HTMLBase plugin included with Eleventy 2.0, which allows
getting rid of virtually all the uses of the "url" filter when referring
to site-local resources. This simplifies content authoring and is less
error prone than the previous approach. Note that a few instances needed
to be replaced with the "htmlBaseUrl" filter included with the plugin.
While on this topic, remove the hardcoded "pathPrefix" option, which in
Eleventy 2.0 always takes precedence over the command line flag, because
the CI setup needs to use the command line flag.