Add XML sitemap and robots.txt for web crawlers (#162)
* Add XML site map

Import a new template that generates an XML site map [1], which web
search engine indexers can use to get a better understanding of the
structure of the site. Pages can specify their change frequency and
the priority indexers should give them, for which the template has
defaults and each page can specify their own values.

[1] https://sitemaps.org/

* Add template for a robots.txt file

This is used by web crawlers, e.g. search engine indexers to determine
what to fetch and what not. As a bonus, they can link to an XML sitemap
to better direct the efforts of crawlers.