Fix regex in `topLevel` filter to match global nav links (#172)
In the global navigation, the `currentPage` class is applied to a
link if the link's URL matches `page.url | topLevel`. This `topLevel`
filter was excluding the trailing slash, resulting in no nav item
getting the `currentPage` class (except `/`). Tweaking the regex in
the `topLevel` filter fixes this.

The `ul` in the global navigation was also generating a class name
based on the `topLevel` filter, so changed its `replace_first` filter
to `replace` to accommodate accordingly.