get-wpe.md: newly organized content on how to get WPE
This is a major rewrite of the contents of explore-wpe.md
and code.md (which is unlisted currently and unreachable).
The approach is to first introduce the components that form
WPE (cog, libwpe, etc) and then offer the different
ways to get it, going from the easiest to the most complex method.

For consistency, rename it from "explore WPE" to "Get WPE", since
this is what this page is about. Also make sure that all links
pointing to this page explain what it is about.

Last but no least, I removed links to the yocto webkit layer
from the Learn&Discover page, as this is already offered in the
"Get WPE" page, to avoid redundance.

The "Code" page still needs to be removed, but I will do that
in a subsequent MR when I get to reorganize the contents of
the "Developers" section.