Add link to documentation directory in each release page
Add a button to the documentation directory under /reference for those
releases that have it. A small JS data script adds a "hasDocumentation"
computed data property to each page, which will be true in release pages
for the "wpewebkit" package, from versions 2.23.90 onwards (inclusive).
The version comparison is done reusing the "semver" package. Just in
case there might be some release after that version that does not have
documentation, the "hideDocumentation" data attribute is also honored.

While at it, fix the post date being close tight to the download and
documentation buttons. Instead, promote the <time> element containing
the date to a block, which causes it to occupy its own "line". To avoid
stealing much vertical space for the buttons, a new "btn-l" class is
added, similar to "btn-xl", which produces smaller buttons.