npm: Update to Eleventy 3.x
The main changes are that ESM modules are now supported, so on top of
switching the syntax for imports/exports, "package" is now a keyword and
some of their uses needed to be changed in the configuration file. The
"baseHtmlUrl" filter is now named "htmlBaseUrl".

While at it, the uneeded "arrayLength" filter was removed, in favor of
using .length directly on arrays. Also the unused "recentReleaseNotes"
collection was removed.