blog: Fix pagination permalink expression
Using mathematical operations inside "{{ expression }}" does not work
in all versions of the Liquid template engine, specially with newer
versions as the templating engine attempts to improve compatibility
with the original implementation. Instead of "+ 1", use the "plus:"
filter to adjust the page number of blog pagination permalinks, and
quote the value in the front matter to avoid the YAML parser choking
up on the colon.