wpt-status: replace the header link to github/wpt with one to the wpt.fyi results
This replaces the link in the header (next to the version selector dropdown),
which previously pointed to the GitHub WPT commit used for the run.
In reality, that link isn't very interesting, since it just points to a
random WPT commit where the run happened.

Instead, we now link to the corresponding results page on wpt.fyi, which
makes it easier to view the actual detailed test output and seems more useful
than opening a random commit on the WPT repository.

The GitHub WPT commit can still be accessed easily from the wpt.fyi results
page by clicking on the WPT version label (with the small GitHub icon),
so it still remains easily accessible.