Fix Savant logos so @2x is an exact ratio, add srcset
Add "srcset" attribute to specify the non-default scaled version of the
Savant logo the success story. While at it, ensure that the 2x version
has a size (540x142) which is exactly the double of the 1x one (270x71)
to avoid odd glitches, and massage the PNGs through oxipng to optimize
them.