wpt-status: Add an alternative view based on results from BCD tests.
Browser Compatibility Data (BCD) test results show a very detailed
list of supported Web APIs, JavaScript features and CSS properties
that can be useful for Web developers looking if they can use a
specific feature in a given WPE version.

This patch reworks the "WPE Specification Support" webpage to offer
this second source of data as an alternative view. It also renames
the page to "WPE Web Support Status" in order to make the title
better suited to include both the BCD and WPT data.

On top of that it adds a few extra features to the web-page:

- Generate and use URL query params, that way is possible to share
  links that open a specific view.
- Cache the json data so it doesn't re-download the jsons each time
  a different version or view is selected.
- Add a search/filter field in the table header that allows to easily
  filter the rows.